// Types/Forms/Fields.ts
export type Fields = {
  login: {
    email: {
      label: string;
      name: string;
      type: string;
      errorMessage: string;
      icon: JSX.Element;
    },
    password: {
      label: string;
      name: string;
      type: string;
      errorMessage: string;
      icon: JSX.Element;
    },
  },
  register: {
    username: {
      label: string;
      name: string;
      type: string;
      errorMessage: string;
      icon: JSX.Element;
    },
    email: {
      label: string;
      name: string;
      type: string;
      errorMessage: string;
      icon: JSX.Element;
    },
    password: {
      label: string;
      name: string;
      type: string;
      errorMessage: string;
      icon: JSX.Element;
    },
  },
};